rename files and update readme
