+++ README.md
@@ -14,3 +14,14 @@ It will take the serial data and transform it to translation and rotation inside
 * Copy the serial folder, it is under PySerial\pyserial-3.2.1\serial (the version can change)
 * Paste the folder in your blender modules folder, 2.83\scripts\modules,
 and now you should be able to use import serial
+
+
+## Tests
+### [Blender Minimal](tests/blender_minimal.py)
+This is just a simpel small Blender add-on to prove that we can read fast enough
+
+### [Read Simulator](tests/read_simulator.py)
+This is a simple python program that reads from serial. Useful for testing and you don't want to do it through Blender add-on
+
+### [Write Simulator](tests/write_similator.py)
+This is a simple python program that writes to serial. Useful for testing and you don't have you Arduino hardware
